+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for myapp
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.0.0 -- 2025-05-06
 
 * First version. Released on an unsuspecting world.
+* Organized the effects system into the AppEffects directory
+* Implemented DB effect for database access
+* Created proper separation between generic effect code and application code
+* Moved all utility modules to AppEffects directory